Make test finding code more reloadable
Two intertwined changes here. First, the total removal of the
Reflections library. It was interfering with attempts to make this
package usable for re-running jUnit/Java tests inside the same VM.

Second, specifically use the Clojure DynamicClassLoader for loading
every Java class defined in the current sources. This ensures that as
long as we have a strategy for changing the bytecode for that class
inside the DynamicClassLoader (and we do ;), these classes will be
successfully reloaded.

Lastly, this means that we need to manually do the reflecting to find
classes that contain jUnit tests.